chore: clean up internals
