+++ packages/comark/src/index.ts
@@ -7,9 +7,6 @@ export {
 export type { AutoCloseOptions, LinkMode } from './internal/parse/auto-close/index.ts'
 
 // Re-export parse utilities
-export { applyAutoUnwrap } from './internal/parse/auto-unwrap.ts'
-
-// Re-export parse utilities
 export * from './parse.ts'
 
 // Re-export the ambient renderer context for live updates
