+++ packages/comark/src/parse.ts
@@ -18,13 +18,12 @@ import taskList from './plugins/task-list.ts'
 import alert from './plugins/alert.ts'
 import html from './plugins/html.ts'
 import frontmatterPlugin from './plugins/frontmatter.ts'
-import { applyAutoUnwrap } from './internal/parse/auto-unwrap.ts'
+import { applyAutoUnwrap, extractReusableNodes } from './internal/parse/utils.ts'
 import { applyUnwrap, resolveUnwrapTags } from './internal/parse/unwrap.ts'
 import { autoCloseMarkdown } from './internal/parse/auto-close/index.ts'
-import { extractReusableNodes } from './internal/parse/incremental.ts'
 import { createSerializedTask, dedupePlugins } from './utils/helpers.ts'
 import { noopTracer, withSpan } from './utils/trace.ts'
-import { tokenListToTree } from './internal/parse/tokenListToTree.ts'
+import { tokenListToTree } from './internal/parse/tree.ts'
 
 // Re-export frontmatter utilities
 export { parseFrontmatter } from './internal/frontmatter.ts'
